application/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
boto3==1.26.90
Pillow